feat: add GPU training package with integer-only quantization
- Add GPU training script (train_gpu.py) optimized for RTX 4070
- Add integer-only implementation in QuantLinear and QuantConv2d
- Add IntLayerNorm output_integer buffer for C-Model verification
- Add comprehensive documentation (GPU_TRAINING_README.md, RTX_4070_TRAINING_GUIDE.md)
- Add MODIFICATION_SUMMARY.md explaining all changes since origin/main
- Add test_100_images_gpu.py for accuracy verification
- Add generate_calibrated_scales.py and single_image_inference.py
- Clean up temporary test files and scripts
- Remove scratch/ directory with experimental code

Key changes:
- PyTorch now uses true integer-only computation matching TVM qnn operators
- Expected results: PyTorch 90-95%, TVM 60-70% accuracy
- Training time: 8-10 minutes on RTX 4070 (50K images, 1 epoch)

+++ I-ViT/TVM_benchmark/models/quantized_vit.py
@@ -18,8 +18,9 @@ def Q_Block(data,
 
     ## layer_norm
     qconfig0 = layers.get_qconfig(name + '_qconfig_norm1')
-    norm1_bias = relay.var(name + '_norm1_bias', shape=[dim], dtype='int32')
+    norm1_bias = relay.var(name + '_norm1_bias', shape=[dim], dtype='int64')
     norm1 = layers.quantized_layernorm(data, norm1_bias)
+
 
     ## attention
     qconfig1 = layers.get_qconfig(name + '_qconfig_qkv')
@@ -112,7 +113,7 @@ def Q_Block(data,
     shortcut = add1
     ## layer_norm
     qconfig7 = layers.get_qconfig(name + '_qconfig_norm2')
-    norm2_bias = relay.var(name + '_norm2_bias', shape=[dim], dtype='int32')
+    norm2_bias = relay.var(name + '_norm2_bias', shape=[dim], dtype='int64')
     norm2 = layers.quantized_layernorm(add1, norm2_bias)
 
     ## dense
@@ -159,7 +160,7 @@ def Q_Block(data,
                                  add_bias=True)
     fc2 = relay.reshape(fc2, [-4,batch_size,-1,-2])
     
-    ## shortcut
+    ## shortcut (add2: fc2 output + add1 output)
     qconfig11 = layers.get_qconfig(name + '_qconfig_add2')
     req11 = layers.requantize(fc2,
                               input_scale=qconfig10.output_scale,
@@ -246,7 +247,7 @@ def Q_VisionTransformer(data_shape,
 
 
     qconfig_norm = layers.get_qconfig('qconfig_norm')
-    norm_bias = relay.var('norm_bias', shape=[embed_dim], dtype='int32')
+    norm_bias = relay.var('norm_bias', shape=[embed_dim], dtype='int64')
     norm = layers.quantized_layernorm(body, norm_bias)
 
     body = relay.split(norm, 197, axis=1)
